add redux state to AddButton
