+++ client/src/components/ui/AddButton.tsx
@@ -6,73 +6,69 @@ import Button from "@mui/material/Button";
 import Fab from "@mui/material/Fab";
 import CheckIcon from "@mui/icons-material/Check";
 import SaveIcon from "@mui/icons-material/Save";
+import {
+  useAppDispatch,
+  useAppSelector,
+} from "../../store/hooks/redux";
+import { ALCO_CONTENT } from "../../constants/alcoConstants";
+import { setDecimal } from "../../store/reducer/alcoHandlers";
+import { addNewDoseToDB } from "../../store/reducer/http/alcoActions";
 
-export default function CircularIntegration() {
-  const [loading, setLoading] = React.useState(false);
-  const [success, setSuccess] = React.useState(false);
-  const timer =
-    React.useRef<ReturnType<typeof setTimeout>>();
+interface IAddButtonProps {
+  volume: string;
+  percent: string;
+}
+
+export function AddButton({
+  volume,
+  percent,
+}: IAddButtonProps) {
+  const { service, currentDate } = useAppSelector(
+    (state) => state.alcoReducer
+  );
+  const { isError, isLoading } = service;
+  const { currentLang } = useAppSelector(
+    (state) => state.appReducer
+  );
+  const dispatch = useAppDispatch();
 
   const buttonSx = {
-    ...(success && {
-      bgcolor: green[500],
-      "&:hover": {
-        bgcolor: green[700],
-      },
-    }),
+    ...(isError
+      ? {
+          bgcolor: "error",
+        }
+      : {
+          bgcolor: "secondary",
+        }),
   };
-
-  React.useEffect(() => {
-    return () => {
-      clearTimeout(timer.current);
-    };
-  }, []);
 
   const handleButtonClick = () => {
-    if (!loading) {
-      setSuccess(false);
-      setLoading(true);
-      timer.current = setTimeout(() => {
-        setSuccess(true);
-        setLoading(false);
-      }, 2000);
+    if (volume && percent) {
+      const vodka = setDecimal(
+        (Number(volume) * Number(percent) * 2.5) / 100,
+        0
+      );
+      dispatch(
+        addNewDoseToDB({
+          additionVodka: vodka + "",
+          ...currentDate,
+        })
+      );
     }
   };
-
+  //TODO add message from server
   return (
     <Box sx={{ display: "flex", alignItems: "center" }}>
-      <Box sx={{ m: 1, position: "relative" }}>
-        <Fab
-          aria-label='save'
-          color='primary'
-          sx={buttonSx}
-          onClick={handleButtonClick}
-        >
-          {success ? <CheckIcon /> : <SaveIcon />}
-        </Fab>
-        {loading && (
-          <CircularProgress
-            size={68}
-            sx={{
-              color: green[500],
-              position: "absolute",
-              top: -6,
-              left: -6,
-              zIndex: 1,
-            }}
-          />
-        )}
-      </Box>
       <Box sx={{ m: 1, position: "relative" }}>
         <Button
           variant='contained'
           sx={buttonSx}
-          disabled={loading}
+          disabled={isLoading}
           onClick={handleButtonClick}
         >
-          Accept terms
+          {ALCO_CONTENT[currentLang].btnAdd}
         </Button>
-        {loading && (
+        {isLoading && (
           <CircularProgress
             size={24}
             sx={{
