+++ client/src/components/ui/index.ts
@@ -2,3 +2,4 @@ export { Email } from "./Email";
 export { ServerMessage } from "./ServerMessage";
 export { Password } from "./Password";
 export { SendButton } from "./SendButton";
+export { AddButton } from "./AddButton";
